move dtos related to external libraries to the own package
