+++ org-eclipse-che-jdt-ls-extension/jdt.ls.extension.core/src/main/java/org/eclipse/che/jdt/ls/extension/core/internal/externallibrary/ExternalLibrariesChildrenCommand.java
@@ -13,8 +13,8 @@ package org.eclipse.che.jdt.ls.extension.core.internal.externallibrary;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import java.util.List;
-import org.eclipse.che.jdt.ls.extension.api.dto.ExternalLibrariesParameters;
-import org.eclipse.che.jdt.ls.extension.api.dto.JarEntry;
+import org.eclipse.che.jdt.ls.extension.api.dto.externallibrary.ExternalLibrariesParameters;
+import org.eclipse.che.jdt.ls.extension.api.dto.externallibrary.JarEntry;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.lsp4j.jsonrpc.json.adapters.CollectionTypeAdapterFactory;
