+++ org-eclipse-che-jdt-ls-extension/jdt.ls.extension.core/src/main/java/org/eclipse/che/jdt/ls/extension/core/internal/externallibrary/LibraryNavigation.java
@@ -22,9 +22,9 @@ import java.util.List;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import org.apache.commons.io.IOUtils;
-import org.eclipse.che.jdt.ls.extension.api.dto.ClassContent;
-import org.eclipse.che.jdt.ls.extension.api.dto.Jar;
-import org.eclipse.che.jdt.ls.extension.api.dto.JarEntry;
+import org.eclipse.che.jdt.ls.extension.api.dto.externallibrary.ClassContent;
+import org.eclipse.che.jdt.ls.extension.api.dto.externallibrary.Jar;
+import org.eclipse.che.jdt.ls.extension.api.dto.externallibrary.JarEntry;
 import org.eclipse.che.jdt.ls.extension.core.internal.JavaModelUtil;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
